backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
pandas==1.3.3
scikit-learn==0.24.2
PyJWT==1.7.1
python-decouple==3.3
